Configure anonymous (readonly) mode via API config
The anonymous readonly mode is configured via the app configuration of
the .NET API. The client retrieves the configuration setting from the
API along with the auth settings. This removes the need of handling two
separate environment variables.

To share the new configuration in the auth settings across react
components, a custom AuthProviderContext is introduced.